requirements.txt
bcrypt==3.2.0
black==22.3.0
cffi==1.15.0
click==8.1.2
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.35
tomli==2.0.1
typing-extensions==4.2.0
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
